components/api/pyproject.toml
[tool.poetry]
name = "src"
version = "1.0.0"
description = "Flask API Implementation"
authors = ["Kennedy Uche <kennedychukwukauche@gmail.com>"]
classifiers = [
    "DevopsChallenge :: Python :: 3",
]

[tool.poetry.dependencies]
python = "^3.10.0"
Flask  = "3.0.0"
flask-cors = "4.0.0"
requests = "2.31.0"
load_dotenv = "0.1.0"
prometheus-client = "0.17.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
api = "src:main"